INCOMPATIBLE CHANGE: changed behavior of edge retrieval methods
The methods EdgeHandler::edges(), EdgeHandler::inEdges() and EdgeHandler::outEdges()
previously returned a list of edge ids in the specified edge collection. This was
in contrast to the methods' signatures and documentation, which suggested that the
methods would return the edges connected to the specified vertex. However, passing
in a vertex or a direction parameter never had an effect because internally the
methods called a server API which did not take these parameters into account.

With the fix, the parameters are taken into account so only connected edges are
returned as documented. Additionally, the return value of the methods change from
an array of edge ids (i.e. strings) to an array of edge objects.

This is not compatible with the behavior of the methods prior to version 2.6, but
staying compatible with the old methods does not make much sense here, as they did
not do what the documentation suggested.